Trait kernel_hal_unix::devices::NetDriver

pub trait NetDriver: Driver {
    fn get_mac(&self) -> Address { ... }

    fn get_ifname(&self) -> String { ... }

    fn get_ip_addresses(&self) -> Vec<Cidr, Global> { ... }

    fn ipv4_address(&self) -> Option<Address> { ... }

    fn poll(&self, _socketset: &Mutex<Set<'_>>) { ... }

    fn send(&self, _data: &[u8]) -> Option<usize> { ... }

    fn get_arp(&self, _ip: Address) -> Option<Address> { ... }

    fn get_device_cap(&self) -> DeviceCapabilities { ... }
}

Provided methods

fn get_mac(&self) -> Address

fn get_ifname(&self) -> String

fn get_ip_addresses(&self) -> Vec<Cidr, Global>

fn ipv4_address(&self) -> Option<Address>

fn poll(&self, _socketset: &Mutex<Set<'_>>)

fn send(&self, _data: &[u8]) -> Option<usize>

fn get_arp(&self, _ip: Address) -> Option<Address>

fn get_device_cap(&self) -> DeviceCapabilities

Implementations

impl dyn NetDriver + 'static

pub fn is<__T>(&self) -> bool where
__T: NetDriver,

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T>(
self: Box<dyn NetDriver + 'static, Global>
) -> Result<Box<__T, Global>, Box<dyn NetDriver + 'static, Global>> where
__T: NetDriver,

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T>(
self: Rc<dyn NetDriver + 'static>
) -> Result<Rc<__T>, Rc<dyn NetDriver + 'static>> where
__T: NetDriver,

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T>(&self) -> Option<&__T> where
__T: NetDriver,

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T>(&mut self) -> Option<&mut __T> where
__T: NetDriver,

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_arc<__T>(
self: Arc<dyn NetDriver + 'static>
) -> Result<Arc<__T>, Arc<dyn NetDriver + 'static>> where
__T: NetDriver + Any + Send + Sync,

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn’t.