Trait kernel_hal_unix::PageTableTrait

pub trait PageTableTrait: Sync + Send {
    fn map(
        &mut self, 
        _vaddr: usize, 
        _paddr: usize, 
        _flags: MMUFlags
    ) -> Result<(), HalError>;

    fn unmap(&mut self, _vaddr: usize) -> Result<(), HalError>;

    fn protect(
        &mut self, 
        _vaddr: usize, 
        _flags: MMUFlags
    ) -> Result<(), HalError>;

    fn query(&mut self, _vaddr: usize) -> Result<usize, HalError>;

    fn table_phys(&self) -> usize;

    fn map_many(
        &mut self, 
        vaddr: usize, 
        paddrs: &[usize], 
        flags: MMUFlags
    ) -> Result<(), HalError> { ... }

    fn map_cont(
        &mut self, 
        vaddr: usize, 
        paddr: usize, 
        pages: usize, 
        flags: MMUFlags
    ) -> Result<(), HalError> { ... }

    fn unmap_cont(&mut self, vaddr: usize, pages: usize) -> Result<(), HalError> { ... }
}

Required methods

fn map(
&mut self,
_vaddr: usize,
_paddr: usize,
_flags: MMUFlags
) -> Result<(), HalError>

Map the page of vaddr to the frame of paddr with flags.

fn unmap(&mut self, _vaddr: usize) -> Result<(), HalError>

Unmap the page of vaddr.

fn protect(&mut self, _vaddr: usize, _flags: MMUFlags) -> Result<(), HalError>

Change the flags of the page of vaddr.

fn query(&mut self, _vaddr: usize) -> Result<usize, HalError>

Query the physical address which the page of vaddr maps to.

fn table_phys(&self) -> usize

Get the physical address of root page table.

Provided methods

fn map_many(
&mut self,
vaddr: usize,
paddrs: &[usize],
flags: MMUFlags
) -> Result<(), HalError>

fn map_cont(
&mut self,
vaddr: usize,
paddr: usize,
pages: usize,
flags: MMUFlags
) -> Result<(), HalError>

fn unmap_cont(&mut self, vaddr: usize, pages: usize) -> Result<(), HalError>